Organize the python_scripts sturcture to allow use of shared libary
* In Production (the container image), python_scripts/lib is installed
  in a dir $WORKSPACE/python_scripts/lib, and it's added in PYTHONPATH
* In Local dev env, to run the scripts, you could either set PYTHONPATH,
  or use `python3 -m python_scripts.bin.<scripts_name>` to run the scripts

Other Important changes:
* renaming: gather-rpms.py → gather_rpms.py
* extract logging/messaging utils from pulp_utils to common lib/__init__.py
