+++ python_scripts/bin/merge_syft_sbom.py
