+++ python_scripts/bin/select_architectures.py
@@ -1,4 +1,5 @@
 #! /usr/bin/python3
+"""Select and configure architectures for RPM builds."""
 
 import argparse
 import glob
@@ -185,10 +186,10 @@ def _main():
 
     build_architectures = allowed_architectures
     if arches['exclusivearch']:
-        print(f"Limit to ExclusiveArch: {arches["exclusivearch"]}")
+        print(f"Limit to ExclusiveArch: {arches['exclusivearch']}")
         build_architectures &= arches["exclusivearch"]
     if arches['excludearch']:
-        print(f"Avoid ExcludeArch: {arches["excludearch"]}")
+        print(f"Avoid ExcludeArch: {arches['excludearch']}")
         build_architectures -= arches["excludearch"]
     if arches['buildarch'] == set(['noarch']):
         selected_architectures = [random.choice(list(build_architectures))]
